# docs/swagger.json
{
  "schemes": ["http", "https"],
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server for Fiber Base API",
    "title": "Fiber Base API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "paths": {
    "/users": {
      "post": {
        "description": "Create a new user with the provided information",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["users"],
        "summary": "Create a new user",
        "parameters": [
          {
            "description": "User creation data",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CreateUserRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/utils.SuccessResponseModel"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.UserResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Invalid request body or validation error",
            "schema": {
              "$ref": "#/definitions/utils.ErrorResponseModel"
            }
          },
          "409": {
            "description": "Email or username already exists",
            "schema": {
              "$ref": "#/definitions/utils.ErrorResponseModel"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/utils.ErrorResponseModel"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Get a specific user by their ID",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["users"],
        "summary": "Get user by ID",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "User ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/utils.SuccessResponseModel"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.UserResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Invalid user ID",
            "schema": {
              "$ref": "#/definitions/utils.ErrorResponseModel"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/utils.ErrorResponseModel"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/utils.ErrorResponseModel"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "dto.CreateUserRequest": {
      "type": "object",
      "required": ["email", "password", "username"],
      "properties": {
        "avatar": {
          "type": "string",
          "example": "https://example.com/avatar.jpg"
        },
        "email": {
          "type": "string",
          "example": "john.doe@example.com"
        },
        "first_name": {
          "type": "string",
          "maxLength": 50,
          "example": "John"
        },
        "last_name": {
          "type": "string",
          "maxLength": 50,
          "example": "Doe"
        },
        "password": {
          "type": "string",
          "minLength": 6,
          "example": "password123"
        },
        "phone": {
          "type": "string",
          "maxLength": 20,
          "example": "+1234567890"
        },
        "username": {
          "type": "string",
          "maxLength": 50,
          "minLength": 3,
          "example": "john_doe"
        }
      }
    },
    "dto.UserResponse": {
      "type": "object",
      "properties": {
        "avatar": {
          "type": "string",
          "example": "https://example.com/avatar.jpg"
        },
        "created_at": {
          "type": "string",
          "example": "2023-01-01T00:00:00Z"
        },
        "email": {
          "type": "string",
          "example": "john.doe@example.com"
        },
        "first_name": {
          "type": "string",
          "example": "John"
        },
        "id": {
          "type": "string",
          "example": "123e4567-e89b-12d3-a456-426614174000"
        },
        "is_active": {
          "type": "boolean",
          "example": true
        },
        "last_name": {
          "type": "string",
          "example": "Doe"
        },
        "phone": {
          "type": "string",
          "example": "+1234567890"
        },
        "updated_at": {
          "type": "string",
          "example": "2023-01-01T00:00:00Z"
        },
        "username": {
          "type": "string",
          "example": "john_doe"
        }
      }
    },
    "utils.ErrorResponseModel": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Error description"
        },
        "message": {
          "type": "string",
          "example": "Error"
        },
        "success": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "utils.SuccessResponseModel": {
      "type": "object",
      "properties": {
        "data": {},
        "message": {
          "type": "string",
          "example": "Operation completed successfully"
        },
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    }
  }
}